# omd_api.openapi.yaml
openapi: 3.1.0
info:
  title: 餐饮店菜单管理 API | v1
  version: 1.0.0
  description: |
    面向**餐饮店老板端 APP** 的菜单管理接口契约。  
    **本文件仅使用 `GET`（查询）与 `POST`（提交/变更）两种 HTTP 方法。**

    > **文档阅读说明**：各接口下方直接写出路径/Query/Body 参数与返回字段；**公共请求头见第二节**，单个接口不再重复列出。

    ---

    ## 一、环境域名

    | 环境 | Base URL | 说明 |
    |------|----------|------|
    | **生产（prod）** | `https://api.offergo.com/restaurant/v1` | 线上正式环境 |
    | **开发（dev）** | `https://dev-api.offergo.com/restaurant/v1` | 联调、测试环境 |

    ---

    ## 二、统一请求头（全局，各接口默认遵循）

    除标注「无需鉴权」的接口外，均需携带 **Authorization**。所有 **POST** Body 为 JSON。

    | Header 名称 | 必填 | 示例 | 作用 |
    |-------------|------|------|------|
    | `Content-Type` | POST 必填 | `application/json` | 请求体类型 |
    | `Accept` | 建议 | `application/json` | 期望响应为 JSON |
    | `Authorization` | 除登录外必填 | `Bearer eyJhbGciOi...` | 登录后 `data.token` |
    | `X-App-Version` | 建议 | `1.2.0` | 客户端版本号 |
    | `X-Platform` | 建议 | `ios` / `android` | 运行平台 |
    | `X-Device-Id` | 建议 | `A1B2C3D4-...` | 设备唯一标识 |
    | `X-Request-Id` | 建议 | UUID | 请求链路 ID |
    | `Accept-Language` | 可选 | `zh-CN` | 错误文案语言 |

    ---

    ## 三、统一返回格式

    HTTP 状态码表示传输层；**业务成败看 Body 中 `code`（string）**。

    ```json
    { "code": "0", "message": "ok", "data": {} }
    ```

    | 字段 | 类型 | 说明 |
    |------|------|------|
    | `code` | string | `"0"` 成功；非 `"0"` 业务失败 |
    | `message` | string | 提示文案 |
    | `data` | object / array / null | 业务数据 |

...

    ---

    ## 四、接口总览

    ### 4.1 应用配置
    | 方法 | 路径 | 说明 |
...

    ### 4.2 认证
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /stores/{storeId}:
    get:
      tags: [门店]
      summary: 门店详情
      description: |
        | 项 | 值 |
        |----|-----|
        | 请求类型 | GET |
        | 是否鉴权 | **是** |

        **路径参数**

        | 字段 | 类型 | 必填 | 说明 |
        |------|------|------|------|
        | storeId | integer | 是 | 门店 Id |

        **返回 data**：门店对象（字段同 GET /stores 列表元素）
      operationId: getStoreDetail
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 20001
      responses:
        '200':
          content:
            application/json:
              example:
                code: "0"
                message: ok
                data:
                  id: 20001
                  name: 老王牛肉面（人民路店）
                  status: open
                  address: 东京都台东区上野 1-2-3
                  phone: "03-1234-5678"
                  contactName: 王小明
                  businessHours:
                    isOpen24Hours: false
                    openTime: 32400000
                    closeTime: 79200000
                  closedWeekdays: [2]
                  serviceInfo:
                    serviceExpireAt: "2027-12-31 23:59:59"
                    vipLevel: 2
                    referrerId: "80001"
                    additionalPeriod: 30
  /stores/save:
    post:
      tags: [门店]
      summary: 新增或编辑店铺
      description: |
        | 项 | 值 |
        |----|-----|
        | 请求类型 | POST |
        | 是否鉴权 | **是** |

        **Body 参数**

        | 字段 | 类型 | 必填 | 说明 |
        |------|------|------|------|
        | id | integer | 否 | 有值=编辑，无值=新增 |
        | name | string | 是 | 店铺名称 |
        | address | string | 是 | 门店地址 |
        | phone | string | 是 | 门店电话 |
        | contactName | string | 是 | 联系人姓名 |
        | vipLevel | integer | 是 | VIP 等级 |
        | referrerId | string | 否 | 推荐人 Id |
        | status | string | 编辑可选 | open=营业 closed=打烊 rest=休息（新增默认 rest） |
        | businessHours | object | 否 | 营业时间，见 GET /stores |
        | closedWeekdays | integer[] | 否 | 定休日，1=周一 … 7=周日 |

        **返回 data**：保存后的门店对象（字段同 GET /stores 列表元素）
      operationId: saveStore
      requestBody:
        required: true
        content:
          application/json:
            examples:
              create:
                value:
                  name: 老王牛肉面（上野店）
                  address: 东京都台东区上野 4-5-6
                  phone: "03-8765-4321"
                  contactName: 王小明
                  vipLevel: 1
                  referrerId: "80001"
                  businessHours:
                    isOpen24Hours: true
                  closedWeekdays: [1, 7]
              update:
                value:
                  id: 20001
                  name: 老王牛肉面（人民路店）
                  address: 东京都台东区上野 1-2-3
                  phone: "03-1234-5678"
                  contactName: 王小明
                  vipLevel: 2
                  status: open
                  businessHours:
                    isOpen24Hours: false
                    openTime: 32400000
                    closeTime: 79200000
                  closedWeekdays: [2]
      responses:
        '200':
          content:
            application/json:
              example:
                code: "0"
                message: ok
                data:
                  id: 20002
                  name: 老王牛肉面（上野店）
                  status: rest
                  address: 东京都台东区上野 4-5-6
                  phone: "03-8765-4321"
                  contactName: 王小明
                  serviceInfo:
                    serviceExpireAt: "2027-06-30 23:59:59"
                    vipLevel: 1
                    referrerId: "80001"
                    additionalPeriod: 0
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 见第二节 Authorization；Token 来自 POST /auth/login 的 data.token